requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==4.2.7
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.3
djangorestframework==3.14.0
idna==3.11
psycopg2-binary==2.9.7
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
